Tarea de usuarios terminadas

+++ src/app/componentes/administracion/admin-usuarios/admin-usuarios.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { UsuarioService } from '../../../services/usuario.service';
-import { faTrashAlt,faPenAlt,faExclamationTriangle,faUnlock,faCamera } from '@fortawesome/free-solid-svg-icons'
+import { faTrashAlt,faPenAlt,faExclamationTriangle,faUnlock,faCamera,faExchangeAlt } from '@fortawesome/free-solid-svg-icons'
 import Swal from 'sweetalert2/src/sweetalert2.js'
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { FormControl,FormGroup,Validators} from '@angular/forms';
@@ -28,6 +28,7 @@ export class AdminUsuariosComponent implements OnInit {
   faExclamation = faExclamationTriangle;
   faUnlock = faUnlock;
   faCamera = faCamera;
+  faExchangeAlt = faExchangeAlt;
 
   constructor(private router:Router,private modalService:NgbModal,private usuarioService:UsuarioService,private firestorage:AngularFireStorage) { }
 
@@ -85,6 +86,16 @@ export class AdminUsuariosComponent implements OnInit {
     telefono  : new FormControl('',[Validators.required,Validators.minLength(8)]),
   });
 
+  RegistroUsuario = new FormGroup({
+    usuario : new FormControl('',[Validators.required,Validators.minLength(3),Validators.maxLength(15)]),
+    nombres : new FormControl('',[Validators.required]),
+    apellidos : new FormControl('',[Validators.required]),
+    correo: new FormControl('',[Validators.required,Validators.email]),
+    contrasena: new FormControl('',[Validators.required]),
+    direccion: new FormControl('',[Validators.required,Validators.minLength(10)]),
+    telefono: new FormControl('',[Validators.required,Validators.minLength(8),Validators.maxLength(10),Validators.pattern(/^-?(0|[1-9]\d*)?$/)])
+  });
+
   get correoP(){
     return this.EditarInfoUsuario.get('correo');
   }
@@ -101,6 +112,70 @@ export class AdminUsuariosComponent implements OnInit {
     return this.EditarInfoUsuario.get('direccion');
   }
 
+  mostrarContrasena(){
+    let elemento :any = document.getElementById('pass');
+    if(elemento.type =='text'){
+      elemento.type = "password";
+    }else{
+      elemento.type = "text";
+    }
+    
+  }
+
+  //////////////////////
+  registrar():void{
+    this.usuarioService.registrarUsuario(this.RegistroUsuario.value).subscribe(res=>{
+      if(res.code){
+        this.Toast.fire({
+          icon: 'error',
+          title: `Este correo ya fue registrado`
+        })
+      }else{
+
+        this.Toast.fire({
+          icon: 'success',
+          title: `Usuario Registrado`
+        })
+        this.usuarioService.obtenerUsuariosPorRol(this.rolUsuarioRegistrado).subscribe(res=>{
+          this.usuariosRegistrados = res.usuarios;
+        });
+        this.RegistroUsuario.reset();
+      }
+    });
+  }
+
+
+  //gets del formGroup
+
+  get usuario(){
+    return this.RegistroUsuario.get('usuario');
+  }
+
+  get nombres(){
+    return this.RegistroUsuario.get('nombres');
+  }
+
+  get apellidos(){
+    return this.RegistroUsuario.get('apellidos');
+  }
+
+  get correo(){
+    return this.RegistroUsuario.get('correo');
+  }
+
+  get contrasena(){
+    return this.RegistroUsuario.get('contrasena');
+  }
+
+  get direccion(){
+    return this.RegistroUsuario.get('direccion');
+  }
+
+  get telefono(){
+    return this.RegistroUsuario.get('telefono');
+  }
+
+
   prueba():void{
     console.log(this.roles);
   }
@@ -115,24 +190,51 @@ export class AdminUsuariosComponent implements OnInit {
   }
 
   eliminarUsuario(idUsuario):void{
-    this.usuarioService.eliminarUsuario(idUsuario).subscribe(res=>{
-      if(res.resultado.n==1){
-        this.Toast.fire({
-          icon: 'success',
-          title: `${res.mensaje}`
-        })
-        this.usuarioService.obtenerUsuariosPorRol(this.rolUsuarioRegistrado)
-        .subscribe(result=>{
-          this.usuariosRegistrados=result.usuarios;
+
+    this.usuarioService.obtenerUsuario(idUsuario).subscribe(respuesta=>{
+      console.log(respuesta);
+      if(respuesta.rol==this.rolUsuarioRegistrado){
+        console.log("ENTRO A REGISTRO")
+        this.usuarioService.eliminarUsuario(idUsuario).subscribe(res=>{
+          if(res.resultado.n==1){
+             this.usuarioService.obtenerUsuariosPorRol(this.rolUsuarioRegistrado)
+             .subscribe(result=>{
+               this.Toast.fire({
+                 icon: 'success',
+                 title: `${res.mensaje}`
+               })
+               this.usuariosRegistrados=result.usuarios;
+            });
+          }else{
+            this.Toast.fire({
+              icon: 'error',
+              title: `${res.mensaje}`
+            })
+          }
         });
-      }else{
-        this.Toast.fire({
-          icon: 'error',
-          title: `${res.mensaje}`
-        })
+      }
+      
+      if (respuesta.rol== this.rolUsuarioAdmin) {
+        console.log("ENTRO A ADMIN");
+        this.usuarioService.eliminarUsuario(idUsuario).subscribe(res=>{
+          if(res.resultado.n==1){
+            this.usuarioService.obtenerUsuariosPorRol(this.rolUsuarioAdmin)
+            .subscribe(result=>{
+              this.Toast.fire({
+                icon: 'success',
+                title: `${res.mensaje}`
+              })
+              this.usuariosAdmins=result.usuarios;
+            });
+          }else{
+            this.Toast.fire({
+              icon: 'error',
+              title: `${res.mensaje}`
+            })
+          }
+        });
       }
     });
-    console.log(idUsuario);
   }
 
   infoUsuarioSeleccionado(idUsuario):void{
@@ -164,6 +266,14 @@ export class AdminUsuariosComponent implements OnInit {
           this.modalService.dismissAll();
           this.usuariosRegistrados=result.usuarios;
         });
+        this.usuarioService.obtenerUsuariosPorRol(this.rolUsuarioAdmin).subscribe(result=>{
+          this.Toast.fire({
+            icon: 'success',
+            title: 'Se edito correctamente'
+          })
+          this.modalService.dismissAll();
+          this.usuariosAdmins=result.usuarios;
+        });
       }
     })
   }
@@ -194,18 +304,84 @@ export class AdminUsuariosComponent implements OnInit {
        this.usuarioService.editarFotoPerfilAdmin(this.usuarioSeleccionado,res).subscribe(result=>{
         if(result.n==1){
           this.usuarioService.obtenerUsuariosPorRol(this.rolUsuarioRegistrado).subscribe(resultado=>{
-            this.Toast.fire({
-              icon: 'success',
-              title: 'Se edito correctamente'
-            })
-            this.modalService.dismissAll();
             this.usuariosRegistrados=resultado.usuarios;
           });
+          this.usuarioService.obtenerUsuariosPorRol(this.rolUsuarioAdmin).subscribe(resultado=>{
+            this.usuariosAdmins=resultado.usuarios;
+          });
+          this.Toast.fire({
+            icon: 'success',
+            title: 'Se edito correctamente'
+          })
+          this.modalService.dismissAll();
+          this.file = null;
+          this.filePath = null;
+          this.fotoSelect = null;
         }
        })
     })
   }
+  
+  cambiarRol(){
+    let usuarioSeleccionado = this.getUsuarioSeleccionado();
+    this.usuarioService.obtenerUsuario(usuarioSeleccionado).subscribe(respuesta=>{
+      console.log(respuesta);
+      if(respuesta.rol==this.rolUsuarioRegistrado){
+        console.log("ENTRO A REGISTRO")
+        this.usuarioService.cambiarRol(usuarioSeleccionado,this.rolUsuarioAdmin).subscribe(res=>{
+          console.log(res)
+          if(res.n==1){
+            this.usuarioService.obtenerUsuariosPorRol(this.rolUsuarioRegistrado)
+            .subscribe(result=>{
+              this.usuariosRegistrados=result.usuarios;
+           });
+           this.usuarioService.obtenerUsuariosPorRol(this.rolUsuarioAdmin)
+            .subscribe(result=>{
+              this.usuariosAdmins=result.usuarios;
+           });
+           this.Toast.fire({
+            icon: 'success',
+            title: `Accion Exitosa`
+          })
+          this.modalService.dismissAll();
+         }else{
+           this.Toast.fire({
+             icon: 'error',
+             title: `Accion Denegada`
+           })
+         }
+        });
+      }
+      
+      if (respuesta.rol== this.rolUsuarioAdmin) {
+        console.log("ENTRO A ADMIN");
+        this.usuarioService.cambiarRol(usuarioSeleccionado,this.rolUsuarioRegistrado).subscribe(res=>{
+          console.log(res)
+          if(res.n==1){
+            this.usuarioService.obtenerUsuariosPorRol(this.rolUsuarioRegistrado)
+            .subscribe(result=>{
+              this.usuariosRegistrados=result.usuarios;
+           });
+           this.usuarioService.obtenerUsuariosPorRol(this.rolUsuarioAdmin)
+            .subscribe(result=>{
+              this.usuariosAdmins=result.usuarios;
+           });
+           this.Toast.fire({
+            icon: 'success',
+            title: `Accion Exitosa`
+          })
+          this.modalService.dismissAll();
+         }else{
+           this.Toast.fire({
+             icon: 'error',
+             title: `Accion Denegada`
+           })
+         }
+        });
+      }
+    });
     
+  }
